tests/requirements.txt
pytest
requests
openai
pytest-asyncio
aiohttp
python-dotenv
boto3